Tighten up timing computation
The current timing computation has an implicit assumption that the
benchmark ran for 10 seconds when in reality it will be a bit off, by
a few microseconds, if not milliseconds.  Fix this by recording the
start and end times at the entry and end points of each thread and
then use the earliest start and latest end times to get an estimate of
the total time the script iterations ran for.

Also increase the precision of the ops/s rate.